Move audio utilities to speechsdk_example so that the base framework has simplified installation via just pip install.
