+++ setup_batchkit.py
@@ -31,7 +31,7 @@ for line in required:
 # Package specification for batchkit library.
 setup(
     name='batchkit',
-    version='0.9.2-dev0',
+    version='0.9.2-dev1',
     author='Microsoft Azure',
     author_email='andwald@microsoft.com',
     description="Generic batch processing framework for managing the orchestration, dispatch, fault tolerance, and "
