+++ setup_examples.py
@@ -17,7 +17,7 @@ deps = [line for line in required if len(line) > 0 and line[0] != "#"]
 # Package specification that includes every example.
 setup(
     name='batchkit_examples_speechsdk',
-    version='0.9.2-dev0',
+    version='0.9.2-dev1',
     author='Microsoft Azure',
     author_email='andwald@microsoft.com',
     url='https://github.com/microsoft/batch-processing-kit',
